refactor: SpinnerAdapter 생성하여 분리
